refactor(FakeRacingStrategy): Modify fake racing strategy name for test
Before
FakeRandomNumberGenerator

After
FakeRacingStrategy

테스트용 RacingStrategy 클래스 이름을 변경했습니다.
