+++ src/test/java/racingcar/domain/car/CarFactoryTest.java
@@ -1,12 +1,10 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.Car;
-import racingcar.domain.car.CarFactory;
 
 class CarFactoryTest {
